django_project_root/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
cryptography==41.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.3
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-bootstrap4==23.2
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.2
gunicorn==21.2.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
